+++ fetch_data/sparql_data/create_property_popularity_data_observation.py
@@ -6,7 +6,7 @@ from data import get_async_session
 from fetch_data.sparql_data.pull_wikidata import get_sparql_results
 from fetch_data.sparql_data.sparql_queries import PROPERTY_POPULARITY_QUERY
 from fetch_data.utils.get_wikibase import get_wikibase_from_database
-from logger import logger
+from logger.get_logger import logger
 from model.database import (
     WikibasePropertyPopularityCountModel,
     WikibasePropertyPopularityObservationModel,
