+++ fetch_data/utils/fetch_data_from_api.py
@@ -4,13 +4,13 @@ import asyncio
 import json
 import requests
 
-from logger import logger
+from logger.get_logger import logger
 
 
 async def fetch_api_data(url: str) -> dict:
     """Fetch API Data"""
 
-    logger.info(f"Querying {url}")
+    logger.debug(f"Querying {url}")
     result = await asyncio.to_thread(requests.get, url, timeout=10)
     query_data = json.loads(result.content)
     if "error" in query_data:
