+++ logger/get_logger.py
@@ -6,6 +6,30 @@ import os
 
 from config import log_level
 
+DEFAULT_ARGS = {
+    "name",
+    "msg",
+    "args",
+    "levelname",
+    "levelno",
+    "pathname",
+    "filename",
+    "module",
+    "exc_info",
+    "exc_text",
+    "stack_info",
+    "lineno",
+    "funcName",
+    "created",
+    "msecs",
+    "relativeCreated",
+    "thread",
+    "threadName",
+    "processName",
+    "process",
+    "wikibase",
+}
+
 
 class OptionalExtraFormatter(logging.Formatter):
     """Supports an optional `wikibase` extra argument"""
@@ -13,12 +37,19 @@ class OptionalExtraFormatter(logging.Formatter):
     def format(self, record: logging.LogRecord) -> str:
         """Format Record"""
 
+        base_fmt = "%(asctime)s | %(levelname)s | %(message)s"
+        wikibase_fmt = (
+            "%(asctime)s | %(levelname)s | wikibase %(wikibase)d | %(message)s"
+        )
+
         if "wikibase" in record.__dict__.keys():
-            self._style._fmt = "%(asctime)s | %(levelname)s | wikibase %(wikibase)d | %(message)s"  # pylint: disable=protected-access
+            self._style._fmt = wikibase_fmt  # pylint: disable=protected-access
         else:
-            self._style._fmt = (  # pylint: disable=protected-access
-                "%(asctime)s | %(levelname)s | %(message)s"
-            )
+            self._style._fmt = base_fmt  # pylint: disable=protected-access
+
+        for k in record.__dict__.keys():
+            if k not in DEFAULT_ARGS:
+                self._style._fmt += f"\n\t{k}: {record.__dict__.get(k)}"
 
         return super().format(record)
