+++ model/enum/wikibase_log_type_enum.py
@@ -6,7 +6,7 @@ from typing import Optional
 
 import strawberry
 
-from logger import logger
+from logger.get_logger import logger
 
 
 @strawberry.enum
@@ -311,6 +311,6 @@ def compile_log_type(record: dict) -> WikibaseLogType:
     try:
         assert log_type is not None
     except AssertionError as exc:
-        logger.error("LogTypeError", stack_info=True, extra={"log": record})
+        logger.error("LogTypeError", extra={"log": record})
         raise NotImplementedError(record) from exc
     return log_type
